+++ src/behavior/decorator/script_precondition.rs
@@ -23,7 +23,7 @@ use crate::{
 // region:      --- Precondition
 /// The `Precondition` behavior is used to check a scripted condition before
 /// executing its child.
-#[derive(Decorator, Debug, Default)]
+#[derive(Decorator, Default)]
 pub struct Precondition;
 
 #[async_trait::async_trait]